# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Short Url",
    "description": "API for short url",
    "contact": {
      "name": "API Short Url",
      "email": "victormoreiracorrea@gmail.com"
    }
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/short": {
      "post": {
        "tags": [
          "Create short"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "origUrl": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/id/{urlId}": {
      "get": {
        "tags": [
          "Find short"
        ],
        "description": "",
        "parameters": [
          {
            "name": "urlId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/short/{short}": {
      "get": {
        "tags": [
          "Find short"
        ],
        "description": "",
        "parameters": [
          {
            "name": "short",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/date/{date}": {
      "get": {
        "tags": [
          "Find short"
        ],
        "description": "",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "helathResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": ""
        },
        "message": {
          "type": "string",
          "example": ""
        }
      }
    },
    "errorResponse.400": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "400"
        },
        "message": {
          "type": "string",
          "example": "Error parameters"
        }
      }
    },
    "errorResponse.403": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "403"
        },
        "message": {
          "type": "string",
          "example": "Access denied"
        }
      }
    },
    "errorResponse.404": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "404"
        },
        "message": {
          "type": "string",
          "example": "Not found"
        }
      }
    },
    "errorResponse.500": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "500"
        },
        "message": {
          "type": "string",
          "example": "Error internal"
        }
      }
    }
  }
}